JD-ResAIAgent/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML Libraries
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-huggingface==0.0.1
sentence-transformers>=2.6.0
faiss-cpu==1.7.4
openai>=1.6.1,<2.0.0

# Document Processing
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.8

# AWS and Cloud Storage
boto3==1.34.0
aioboto3==12.3.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
uuid==1.30

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# CORS and Middleware
starlette==0.27.0
